bdev/nvme: Add spdk_bdev_get_ctrlr_for_each_subsystem function
This patch adds a new function spdk_bdev_get_ctrlr_for_each_subsystem
which will return all the controllers associated with all the
subsystems in the bdev layer through a callback function.

This function allows the caller to get certain controller specific
properties from the ctrlr object passed in the callback function. One
of the use cases is to get the io_queue_size and io_queue_requests
supported by the controller. Earlier we had no way to know the actual
queue depth supported by the controller as no option of getting the
controller object in the bdev layer was present.

Change-Id: I36dd4a0664dabcbb0b32429725530e60c7526175
Signed-off-by: Yash Raj Singh <yash.rajsingh@nutanix.com>