nvmf: use bdev's nsid for admin command passthru
The ID of a namespace exposed by an NVMe-oF subsystem can differ from
the nsid of the underlying namespace that backs a given bdev.  For
instance, the following:

  $ scripts/rpc.py bdev_nvme_attach_controller -b nvme0 ...
  nvme0n1
  nvme0n2
  $ scripts/rpc.py nvmf_subsystem_add_ns -n 2 nqn nvme0n1
  $ scripts/rpc.py nvmf_subsystem_add_ns -n 1 nqn nvme0n2

results in nvme0n1 being exposed with nsid=2, while nvme0n2 with nsid=1.
In this scenario, admin commands targeting a namespace (e.g. identify
namespace) would get sent to the wrong namespace.

Fixes: 05632afd77dd ("nvmf: large IU and atomic write unit reporting")
Reported-by: Maciej Szulik <maciej.szulik@intel.com>
Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: Ib4f7071afb10691c88def880fadc5f4d7e134b07