lib/nvme: eventfd to handle disconnected I/O qpair
In poll mode the process completion API checks for the disconnected
queue pair and calls the specified disconnected_qpair_cb function. In
interrupt mode we need a way to notify the application in case of
disconnected queue pair.
Register an eventfd to the epoll file descriptor of nvme poll group. This
fd will be written after a queue pair disconnection. Thus the next event
on poll group will handle the queue pair disconnection.

Change-Id: I481eeb847c22e710d8e9705c4ea3e9267da93876
Signed-off-by: Ankit Kumar <ankit.kumar@samsung.com>