correct kick for CQ full case
We implement basic flow control by not processing any submissions while
the CQ is full. However, in interrupt mode, the code was mistakenly
attempting to kick ctrlr->intr_fd - this will not go well, as this is
the vfio-user socket fd, not an eventfd. It should be kicking the SQ
group's eventfd.

Fixes: 23b518a0 ("nvmf/vfio_user: mitigate cq full race")
Signed-off-by: John Levon <john.levon@nutanix.com>
Change-Id: Ia3b5ad8ffe4118a620fc3894e40fc95c251cd009