thread: fd_group-based interrupts
It's now possible to register an interrupt for a whole fd_group.  The
advantage of doing this over registering an interrupt using fd_group's
fd is that the fd_group is nested in thread's fd_group, so
spdk_fd_group_wait() on thread's fd_group will trigger events of the
registered fd_group.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I1b2e4e9ea0b5dc2a8ba5e7ab7366fe1c412167f5