lib/reduce: Add a chunk data read/write cache
This cache can improve the performance of small sequential I/O read and
write operations.

struct reduce_vol_chunk_data_cache is a read/write cache that contains
REDUCE_NUM_VOL_REQUESTS instances of struct reduce_vol_chunk_data, which
is an abstraction of 16KB data blocks. The functions chunk_cache_find_chunk
and chunk_cache_put can be used to retrieve and return cached data to the
cache. Additionally, if the required data is not in the cache, it can be
treated as a memory pool, with memory blocks acquired and returned using
the functions chunk_cache_get_free and chunk_cache_put_free.

Change-Id: Ieb76a0ad1e51e93734aac89b8f5fd45167c29bcf
Signed-off-by: Yankun Li <yankun@staff.sina.com>