nvme: add poll_group interrupt callback
In interrupt mode, IO completions are processed when waiting on
poll_group's fd_group.  But there are some events (qpair disconnection)
that require extra handling.  Normally, this happens in
spdk_nvme_poll_group_wait(), but when manually doing a
spdk_fd_group_wait() on poll_group's fd_group, we need a notification to
get this done.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I08979e42ff57b53f0c97670e9996b0ce6dad713e