lib/nvmf: Fix double free of connect request
When a duplicated qid is detected, we store the connect request to retry
the connect later (in case the other queue with the same qid was in a
disconnection process).
If we successfully create the qpair on second try, we should clear the
connect_req, otherwise it may be confused with a first_fused_req
(because of the union).
When we receive an IO later, we think that there was a previous fused
command, so we abort the fused command (i.e. the connect request). So
this command is freed twice.

Change-Id: I66d0e6c813c2b4378bf03ebac68914340737632c
Signed-off-by: Alex Michon <amichon@kalrayinc.com>