lib/reduce: Support storing metadata on the backing dev.
When creating a compress device, if we don't key in PM_PATH parameter, the metadata will be stored at the beginning of the backing device. This enhances the performance and reliability of compression on devices without persistent memory, primarily reflected in the following ways:
1.It eliminates the need to use nearly 0.3% of the logical disk’s capacity for memory mapping with physical memory.
2.It avoids the need to call the time-consuming 'pmem_msync' synchronization interface, thereby reducing CPU usage.
3.It removes concerns about data loss caused by damage to the file system medium where the metadata resides.

Additionally, it provides a pass-through write cache function for metadata and a metadata merge writing function, further enhancing metadata performance and reducing the number of metadata write operations.

Change-Id: I14fd3ce43943c42a11b2d1cd9d3cbd8c248b84af
Signed-off-by: Yalong Wang <yalong9@staff.sina.com.cn>