lib/reduce: Unmap backing dev blocks
Return the free blocks in the backing bdev to the underlying layer (lvs)
to keep the backing bdev thin.
This patch will send unmap operations to the backing bdev to release the
corresponding blocks. The modification is divided into two phases:

Phase 1 - Adjust the unmap processing logic in the reduce function and use
req->unmap_iov instead of calloc.

Change-Id: I39544e2d9d8e23a69d95508446d743b4f726229c
Signed-off-by: Yankun Li <yankun@staff.sina.com>