lib/reduce: Unmap backing dev blocks
Return the free blocks in the backing bdev to the underlying layer (lvs)
to keep the backing bdev thin.
This patch will send unmap operations to the backing bdev to release the
corresponding blocks. The modification is divided into two phases:

Phase 2 - Calculate the range of the backing device that corresponds to the
free io_unit in the unmap operation. This range must consist entirely of
free io_units and should be of the size recommended by the backing device.
Then, build an unmap operation based on this range.

Change-Id: I041b18f19d1e7524875b8cb93bf0d5c87574783d
Signed-off-by: Yankun Li <yankun@staff.sina.com>