lib/reduce: Send unmap to backing dev
This patch will calculate the range of the backing device that
corresponds to the free io_unit in the unmap operation. This range
must consist entirely of free io_units and should be of the size
recommended by the backing device. Then, build an unmap operation
based on this range.

Multiple released io units in one chunk may be within the same optimal io
boundary. In this case, we only launch an unmap operation on the range.

Change-Id: I041b18f19d1e7524875b8cb93bf0d5c87574783d
Signed-off-by: Yankun Li <yankun@staff.sina.com>