lib/reduce: Support storing metadata on backing dev. (1 of 6, base of metablock cache/operation)
Provides a method for achieving higher performance CompressBdev without using PMem.

When creating a compress device, if we don't key in PM_PATH parameter, the metadata will be stored at the beginning of the backing device. This enhances the performance and reliability of compression on devices without persistent memory, primarily reflected in the following ways:
1.It eliminates the need to use nearly 0.3% of the logical disk’s capacity for memory mapping with physical memory.
2.It avoids the need to call the time-consuming 'pmem_msync' synchronization interface, thereby reducing CPU usage.
3.It removes concerns about data loss caused by damage to the file system medium where the metadata resides.

Additionally, it provides a write-through write cache function for metadata and a metadata merge writing function, further enhancing metadata performance and reducing the number of metadata write operations.

Change-Id: I1c448263cadcf4ac6ab4262a58e640bd70ffe0dd
Signed-off-by: Yalong Wang <yalong9@staff.sina.com.cn>