test/check_so_deps: use VERSION to look for prior tags
The check_so_deps script compares the ABI of the current commit from the
ABI of the previous release.  It used to find the previous release by
looking at the first non-LTS, non-*pre tag and extracting the version
from that tag.  This means that after any other tag is placed, it'll be
used to figure out the release to test against.

Instead, use VERSION to find the release tag to compare against.  That
way, as long as VERSION points to a -pre tag, we can place any number of
-rc tags without breaking the ABI tests.  And, once a release branch is
created, VERSION would get updated to the -rc tag that points to the
first commit on that branch, so the subsequent commits would be compared
to the ABI of that release.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I1b4f9118e6d3749e0c43da3fa917c08868796eb9