mk/spdk.common.mk Use pattern substitution instead of prefix removal
Older versions of make can't handle the ${v#0} nicely as they don't
treat '#' as part of the $(shell ...) cmd. This leads to:

spdk/mk/spdk.common.mk:584: *** unterminated call to function 'shell': missing ')'.  Stop.

We can't escape that char (it would work under older make) as $${v\#0}
will break under newer versions of make.

So instead, just remove the 0 by replacing it with '' ONLY if it is
an actual MM prefix.

Change-Id: Iaac6fb66ea1addb139e53b1f4cf966e7d5a4745d
Signed-off-by: Michal Berger <michal.berger@intel.com>