mk/spdk.common.mk sed the minor version
Older versions of make can't handle shell's param expansion nicely
as in case of something like ${v#0}, the '#' is not treated as part
of the $(shell ...) cmd. This leads to:

spdk/mk/spdk.common.mk:584: *** unterminated call to function 'shell': missing ')'.  Stop.

Escaping \# won't work for newer make builds and some shells that make
may use are still limitted when it comes to other param expansion
variants.

So to make sure it works in all places, just delegate the main work
to sed.

Change-Id: Iaac6fb66ea1addb139e53b1f4cf966e7d5a4745d
Signed-off-by: Michal Berger <michal.berger@intel.com>