bdev/nvme: Fix IO deadlock between ctrlr reset and destruct
If we lose the connection to a remote nvme device and we attempt to
destroy the controller at the same time, we may end up in the following
scenario:
1) An IO is sent to the remote nvme device
2) The connection is lost
3) We receive a bdev_nvme_detach_controller on the ctrlr. We set
   `nvme_ctrlr->destruct = true`.
4) Then we wait for all qpairs to be closed (i.e. for users to close
   their IO channels).
5) If the user configured a timeout, the IO times out. We go for a reset
   but we don't do anything because we see that `nvme_ctrlr->destruct`
   is true.
The IO will never complete. The user doesn't close its IO channel
because it is still waiting for a callback for its IO. So the controller
is never actually destroyed. We are stuck.

The proposed solution is to destroy the qpairs if we attempt a reset
while the controller is being destroyed. This should abort all pending
IOs and get us out of this deadlock.

Change-Id: Ied5fcc2ca10a34609ca7ccc274192bb7d0ef536c
Signed-off-by: Alex Michon <amichon@kalrayinc.com>